+++ pkg/manual/config.go
@@ -33,36 +33,6 @@ import (
 	"github.com/spf13/viper"
 )
 
-const (
-	ETH_TX_LIST  = "ETH_TX_LIST"
-	ETH_ADDR_LOG = "ETH_ADDR_LOG"
-
-	defaultGenKeyWritePathPrefix = "./accounts/keys/"
-	defaultAddrLogPath           = "./accounts/addresses/accounts"
-
-	typeSuffix            = ".type"
-	httpPathSuffix        = ".http"
-	toSuffix              = ".to"
-	amountSuffix          = ".amount"
-	gasLimitSuffix        = ".gasLimit"
-	gasPriceSuffix        = ".gasPrice"
-	gasPremiumSuffix      = ".gasPremium"
-	feeCapSuffix          = ".feeCap"
-	dataSuffix            = ".data"
-	senderKeyPathSuffix   = ".senderKeyPath"
-	writeSenderPathSuffix = ".writeSenderPath"
-	l1SenderSuffix        = ".l1Sender"
-	l1RollupTxIdSuffix    = ".l1RollupTxId"
-	sigHashTypeSuffix     = ".sigHashType"
-	frequencySuffix       = ".frequency"
-	totalNumberSuffix     = ".totalNumber"
-	delaySuffix           = ".delay"
-	startingNonceSuffix   = ".startingNonce"
-	queueOriginSuffix     = ".queueOrigin"
-	chainIDSuffix         = ".chainID"
-	contractWriteSuffix   = ".writeDeploymentAddrPath"
-)
-
 // TxParams holds the parameters for a given transaction
 type TxParams struct {
 	// Name of this tx in the .toml file
@@ -111,9 +81,7 @@ type TxParams struct {
 
 // NewConfig returns a new tx spammer config
 func NewTxParams() ([]TxParams, error) {
-	viper.BindEnv("eth.txs", ETH_TX_LIST)
-	viper.BindEnv("eth.addrLogPath", ETH_ADDR_LOG)
-
+	bindEnv()
 	addrLogPath := viper.GetString("eth.addrLogPath")
 	txs := viper.GetStringSlice("eth.txs")
 	txParams := make([]TxParams, len(txs))
